requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
apache-beam==2.46.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==1.2.3
astor==0.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
avro-python3==1.10.2
Babel==2.12.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contextlib2==21.6.0
contourpy==1.0.7
crcmod==1.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.8
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
etils==1.2.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
exif==1.6.0
fastapi==0.95.1
fastavro==1.7.3
fasteners==0.18
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.1
flask-babel==3.1.0
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.4.0
gensim==4.3.1
gevent==22.10.2
geventhttpclient==2.0.2
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
hdfs==2.7.0
httpcore==0.17.0
httplib2==0.21.0
httpx==0.24.0
huggingface-hub==0.14.1
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.27.0
imgaug==0.4.0
immutabledict==2.2.4
importlib-metadata @ file:///C:/b/abs_20ndzb2j6v/croot/importlib-metadata_1678997085534/work
importlib-resources==5.12.0
imutils==0.5.4
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///C:/b/abs_d1yx5tjhli/croot/ipython_1680701887259/work
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labelImg==1.8.6
lazy_loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
lmdb==1.4.1
lvis==0.5.3
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///C:/Users/Mahran/Desktop/Year%203/Assem/Arabic-License-Plate-Recognition/codes/models/research
objsize==0.6.1
onnx==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.10
pandas==2.0.0
pandocfilters==1.5.0
pdf2docx==0.5.6
Pillow==9.5.0
pillow-heif==0.10.1
plum-py==0.8.6
ply==3.11
portalocker==2.7.0
premailer==3.10.0
prometheus-client==0.16.0
promise==2.3
proto-plus==1.22.2
protobuf==3.20.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.17
pydantic==1.10.7
pydot==1.4.2
pydub==0.25.1
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rapidfuzz==3.0.0
rarfile==4.0
regex==2023.3.23
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
sacrebleu==2.2.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentencepiece==0.1.98
seqeval==1.2.2
shapely==2.0.1
sip==6.7.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post4
smart-open @ file:///home/conda/feedstock_root/build_artifacts/smart_open_split_1673202927732/work/dist
sniffio==1.3.0
soupsieve==2.4.1
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.26.1
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.7.4
tensorflow-text==2.10.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1662458743919/work
tqdm==4.65.0
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
tritonclient==2.33.0
typeguard==2.13.3
typing_extensions @ file:///C:/b/abs_a1bb332wcs/croot/typing_extensions_1681939523095/work
tzdata==2023.3
uc-micro-py==1.0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
uvicorn==0.21.1
visualdl==2.5.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.0
widgetsnbextension==4.0.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.15.0
yarl==1.9.2
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
